+++ frontend/src/components/index.js
@@ -1,2 +1,4 @@
 export { Filter } from "./Filter";
 export { Sort } from "./Sort";
+export {Add_to_cart_button} from "./AddToCartButton"
+export {Add_to_wishlist_button} from "./AddToWishlistButton"
