+++ frontend/src/pages/Fertilizer.jsx
@@ -2,9 +2,7 @@ import { useEffect } from "react";
 import { useReduce } from "../Reducer-context/Reducer-context";
 import { Products } from "./Products";
 export function Fertilizer({
-  filteredData,
-  Add_to_cart_button,
-  Add_to_wishlist_button
+  filteredData
 }) {
   let { dispatch } = useReduce();
   useEffect(() => {
@@ -16,8 +14,6 @@ export function Fertilizer({
     <>
       <Products
         filteredData={filteredData}
-        Add_to_cart_button={Add_to_cart_button}
-        Add_to_wishlist_button={Add_to_wishlist_button}
       />
     </>
   );
