+++ frontend/src/pages/ProductPage.jsx
@@ -1,12 +1,15 @@
 import { useParams } from "react-router-dom";
 import { useReduce } from "../Reducer-context/Reducer-context";
-import { Link, useNavigate } from "react-router-dom";
+import { useLogin } from "../Reducer-context/LoginContext";
+import { Link, useNavigate,useLocation } from "react-router-dom";
 import { useEffect } from "react";
-
-export function ProductPage({ Add_to_cart_button, Add_to_wishlist_button }) {
+import { Add_to_cart_button, Add_to_wishlist_button } from "../components";
+export function ProductPage() {
   const { productid } = useParams();
-  const { data } = useReduce();
+  const { data,dispatch,cartlist, wishlist } = useReduce();
   let navigate = useNavigate();
+  const {pathname}=useLocation();
+  const {isUserLogIn}=useLogin()
   let productobj;
   if (data.length > 0) {
     productobj = data.find((item) => item._id === productid);
@@ -22,7 +25,7 @@ export function ProductPage({ Add_to_cart_button, Add_to_wishlist_button }) {
             <div className="productpage-image">
               <img src={productobj.image} alt="productimg" />
               <div className="productpage-addwish">
-                {Add_to_wishlist_button(productobj._id)}
+                {Add_to_wishlist_button(productobj._id,wishlist,dispatch,isUserLogIn,navigate,pathname)}
               </div>
             </div>
 
@@ -53,7 +56,7 @@ export function ProductPage({ Add_to_cart_button, Add_to_wishlist_button }) {
               )}
               <div className="productpage-action">
                 <div className="productpage-addcart">
-                  {Add_to_cart_button(productobj._id, productobj.inStock)}
+                  {Add_to_cart_button(productobj._id, productobj.inStock,cartlist,dispatch,isUserLogIn,navigate,pathname)}
                 </div>
               </div>
             </div>
@@ -121,8 +124,8 @@ export function ProductPage({ Add_to_cart_button, Add_to_wishlist_button }) {
                             <div class="price">
                               Rs. {(price - (price * offer) / 100).toFixed(2)}
                             </div>
-                            {Add_to_cart_button(_id, inStock)}
-                            {Add_to_wishlist_button(_id)}
+                            {Add_to_cart_button(_id, inStock,cartlist,dispatch,isUserLogIn,navigate,pathname)}
+                            {Add_to_wishlist_button(_id, wishlist,dispatch,isUserLogIn,navigate,pathname)}
                           </div>
                         </Link>
                       </>
